fix : security config , redis serialization, deserialization
