+++ user/src/main/java/com/sparta/hotsix/user/UserApplication.java
@@ -1,11 +1,14 @@
 package com.sparta.hotsix.user;
 
+import org.hibernate.annotations.QueryCacheLayout;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+
+@QueryCacheLayout
 public class UserApplication {
 
     public static void main(String[] args) {
