+++ user/src/main/java/com/sparta/hotsix/user/config/appConfig.java
@@ -5,6 +5,9 @@ package com.sparta.hotsix.user.config;
 //import io.swagger.v3.oas.models.security.SecurityRequirement;
 //import io.swagger.v3.oas.models.security.SecurityScheme;
 //import org.springframework.cache.CacheManager;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.security.SecurityRequirement;
@@ -18,9 +21,11 @@ import org.springframework.data.redis.cache.CacheKeyPrefix;
 import org.springframework.data.redis.cache.RedisCacheConfiguration;
 import org.springframework.data.redis.cache.RedisCacheManager;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.RedisSerializationContext;
 import org.springframework.data.redis.serializer.RedisSerializer;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
@@ -28,29 +33,18 @@ import java.time.Duration;
 
 @Configuration
 @EnableJpaAuditing
-@EnableCaching
 public class appConfig {
     //JpaAuditing
     @Bean
-    public AuditorAware<Long> auditorAware(){
+    public AuditorAware<Long> auditorAware() {
         return new CustomAuditorAware();
-    };
+    }
+
+    ;
 
     @Bean
     public PasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public RedisCacheManager cacheManager(RedisConnectionFactory connectionFactory) {
-        RedisCacheConfiguration config = RedisCacheConfiguration
-                .defaultCacheConfig()
-                .disableCachingNullValues()
-                .entryTtl(Duration.ofHours(1))
-                .computePrefixWith(CacheKeyPrefix.simple())
-                .serializeValuesWith(
-                        RedisSerializationContext.SerializationPair.fromSerializer(new Jackson2JsonRedisSerializer<>(Object.class))                );
-    return RedisCacheManager.builder(connectionFactory).cacheDefaults(config).build();
     }
