+++ user/src/main/java/com/sparta/hotsix/user/dto/UserResponse.java
@@ -1,15 +1,20 @@
 package com.sparta.hotsix.user.dto;
 
-import com.sparta.hotsix.user.domain.UserRole;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
+import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.io.Serializable;
 import java.time.LocalDateTime;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserResponse {
+public class UserResponse implements Serializable {
 
     private Long id;
 
@@ -23,9 +28,11 @@ public class UserResponse {
 
     private LocalDateTime createdAt;
 
+
     private LocalDateTime updatedAt;
 
     private String updatedBy;
+
 
     private LocalDateTime deletedAt;
