+++ user/src/main/resources/application.yml
@@ -3,7 +3,7 @@ spring:
     name: user-service
   jpa:
     hibernate:
-      ddl-auto: create
+      ddl-auto: update
 
 
   datasource:
